+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support variadic setter function. [PR#68](https://github.com/JsonMapper/JsonMapper/pull/68)
 
 ## [2.0.0] - 2021-01-07
 ### Changed